pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cesiumjs-anywidget"
version = "0.6.0"
description = "A Jupyter widget for CesiumJS 3D globe visualization using anywidget"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Alexis Placet" }
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Jupyter",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anywidget>=0.9.0",
    "exifread>=3.5.1",
    "ipycanvas>=0.14.3",
    "jupyterlab>=4.3.8",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "pygeodesy>=25.11.5",
    "scipy>=1.10.1",
    "traitlets>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.0.0",
    "notebook>=7.0.0",
    "piexif>=1.1.3",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/Alex-PLACET/cesiumjs_anywidget"
Repository = "https://github.com/Alex-PLACET/cesiumjs_anywidget"

[tool.hatch.build.targets.wheel]
packages = ["src/cesiumjs_anywidget"]
exclude = [
    "src/cesiumjs_anywidget/js",
]
force-include = { "src/cesiumjs_anywidget/index.js" = "cesiumjs_anywidget/index.js" }

[tool.hatch.build.targets.sdist]
exclude = [
    "tests",
    "src/cesiumjs_anywidget/js",
    ".github",
    ".vscode",
    "examples",
    "htmlcov",
    ".coverage",
    "**/*.pyc",
    "**/__pycache__",
    "test_*.py",
    "*_test.py",
    "setup_dev.sh",
    "run_tests.sh",
]
force-include = { "src/cesiumjs_anywidget/index.js" = "src/cesiumjs_anywidget/index.js" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=cesiumjs_anywidget",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src/cesiumjs_anywidget"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
